move certification helpers into middleware file
