+++ js/src/redux/providers/certifications/middleware.js
@@ -14,11 +14,55 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
+import { bytesToHex, hex2Ascii } from '../../../api/util/format';
 import { addCertification } from './actions';
-import fetchCertifier from './fetch-certifier';
-import checkIfCertified from './check-if-certified';
 
+import badgeRegABI from '../../../contracts/abi/badgereg.json';
+import certifierABI from '../../../contracts/abi/certifier.json';
+
+const badgeRegAddress = '0xcF5A62987294fd2087252FD812443508528C52bF';
+const ZERO = '0x0000000000000000000000000000000000000000000000000000000000000000';
 const knownCertifiers = [ 'smsverification' ];
+
+export const fetchCertifier = (api) => {
+  const registry = api.newContract(badgeRegABI, badgeRegAddress);
+  const cache = {};
+
+  return (name) => {
+    if (cache[name]) {
+      return Promise.resolve(cache[name]);
+    }
+    return registry.instance.fromName.call({}, [name])
+    .then(([ id, address ]) => {
+      return Promise.all([
+        registry.instance.meta.call({}, [id, 'TITLE']),
+        registry.instance.meta.call({}, [id, 'IMG'])
+      ])
+        .then(([title, img]) => {
+          title = bytesToHex(title);
+          title = title === ZERO ? null : hex2Ascii(title);
+          if (bytesToHex(img) === ZERO) img = null;
+
+          const data = { address, name, title, icon: img };
+          cache[name] = data;
+          return data;
+        });
+    });
+  };
+};
+
+export const checkIfCertified = (api) => {
+  const contracts = {};
+
+  return (certifier, address) => {
+    if (!contracts[certifier]) {
+      contracts[certifier] = api.newContract(certifierABI, certifier);
+    }
+    const contract = contracts[certifier];
+
+    return contract.instance.certified.call({}, [address]);
+  };
+};
 
 export default (api) => {
   const fetch = fetchCertifier(api);
